Feature/rename bitbucket modules (#2)
* refactor: add credentials to test bitbucket modules

* refactor: rename bitbucket repo module

* refactor: rename bitbucket modules
