+++ .devcontainer/.env.sample
@@ -1,0 +1,2 @@
+BB_USER=__change_me__
+BB_PASS=__change_me__
