+++ .devcontainer/docker-compose.yml
@@ -1,6 +1,7 @@
 services:
   devcontainer:
     image: i2btech/ansible-collection
+    env_file: .env
     volumes:
       - ..:/workspace
       - $HOME/.ssh:/home/ubuntu/.ssh
