+++ src/plugins/modules/bitbucket_repo_env.py
@@ -1,6 +1,6 @@
 #!/usr/bin/python
 # -*- coding: utf-8 -*-
-""" repo_var module """
+""" bitbucket_repo_env module """
 
 # Copyright: (c) 2018, Terry Jones <terry.jones@example.org>
 # GNU General Public License v3.0+ (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
@@ -9,11 +9,11 @@ __metaclass__ = type
 
 DOCUMENTATION = r'''
 ---
-module: repo_var
-short_description: Manage variables for a repository on Bitbucket Cloud
+module: bitbucket_repo_env
+short_description: Manage deployment environment for a repository on Bitbucket Cloud
 version_added: "1.0.0"
 description:
-    - Manage variables for a repository on Bitbucket Cloud
+    - Manage deployment environment for a repository on Bitbucket Cloud
 options:
     username:
         description:
@@ -30,6 +30,17 @@ options:
             - Repository name.
         type: str
         required: true
+    name:
+        type: str
+        description:
+            - Name of environment
+        required: true
+    type:
+        type: str
+        description:
+            - Type of deployment environment
+        choices: [ Test, Staging, Production ]
+        required: true
     variables:
         description: List of variables that will be managed
         type: list
@@ -39,19 +50,19 @@ options:
             name:
                 type: str
                 description:
-                - Name of the variable
+                    - Name of the variable
                 required: true
             value:
                 type: str
                 description:
-                - Value of the variable
+                    - Value of the variable
                 required: true
             secured:
                 type: bool
                 description:
                 - If true, variable will be encrypted and masked in the logs
                 - If true, the module will always finish with a state of changed. The Bitbucket Cloud API don't allow to check
-                  the current value of a secured variable thus we need to always update tha value to applies possible changes on the value
+                the current value of a secured variable thus we need to always update tha value to applies possible changes on the value
                 - If false, you need to delete the variable to change it to secured
                 required: false
 author:
@@ -59,17 +70,19 @@ author:
 '''
 
 EXAMPLES = r'''
-- name: "Set variables for repository X"
-  i2btech.ops.repo_var:
+- name: "Set deployment environments for repository X"
+  i2btech.ops.bitbucket_repo_env:
   username: "alice"
   password: "app_password"
   repository: "example-X"
+  name: Integration
+  type: Test
   variables:
     - name: user
-      value: xxx
+      value: user_db
     - name: pass
-      value: yyy
-      secured: True
+      value: _super_secret_pass_
+      secured: true
 '''
 
 RETURN = r'''
@@ -85,18 +98,39 @@ from ansible_collections.i2btech.ops.plugins.module_utils.bitbucket import Bitbu
 from ansible.module_utils.basic import AnsibleModule
 #pylint: disable=wrong-import-position
 
-def manage_variables(result, bitbucket, module):
-    """ CRUD variables """
+def manage_environments(result, bitbucket, module):
+    """ CRUD environments """
 
-    api_url=bitbucket.BITBUCKET_API_ENDPOINTS['repos-pipeline'].format(
-                url=module.params['url'],
-                workspace='i2b',
-                repo_slug=module.params['repository'])
+    current_environments = bitbucket.get_repository_environments()
+    env_exists = False
+    env_uuid = None
+    for x in current_environments:
+        if module.params['name'].lower() == x['name'].lower() and module.params['type'].lower() == x['environment_type']['name'].lower():
+            env_exists = True
+            env_uuid = x['uuid']
+            break
 
-    current_variables = bitbucket.get_variables(api_url + "/variables/")
-    new_variables = []
-    for var in module.params['variables']:
-        new_variables.extend([var])
+    if env_exists:
+        # environment exists, manage variables associated with it if they exists
+        if module.params['variables'] is not None:
+            api_url=bitbucket.BITBUCKET_API_ENDPOINTS['repos-deployments'].format(
+                        url=module.params['url'],
+                        workspace='i2b',
+                        repo_slug=module.params['repository'])
+            current_variables = bitbucket.get_variables(api_url + "/environments/" + env_uuid + "/variables")
+            manage_environment_variables(result, bitbucket, env_uuid, current_variables, module.params['variables'])
+
+    else:
+        # environment doesn't exist on current environments, add it
+        new_env = bitbucket.manage_repository_environments('create', module.params['name'], module.params['type'])
+        env_uuid = new_env['uuid']
+        # manage variables associated with it
+        if module.params['variables'] is not None:
+            manage_environment_variables(result, bitbucket, env_uuid, [], module.params['variables'])
+        result['changed'] = True
+
+def manage_environment_variables(result, bitbucket, env_uuid, current_variables, new_variables):
+    """ CRUD variables of environment """
 
     # actions for new variables
     for i in new_variables:
@@ -114,17 +148,17 @@ def manage_variables(result, bitbucket, module):
         if var_exists:
             # always update secured variable
             if var_value is None:
-                bitbucket.manage_repository_variables('update', i['name'], i['value'], var_uuid, i['secured'])
+                bitbucket.manage_environment_variables('update', i['name'], i['value'], env_uuid, var_uuid, i['secured'])
                 result['changed'] = True
             # if not secured and value are different, update variable
             else:
                 if i['value'] != var_value:
-                    bitbucket.manage_repository_variables('update', i['name'], i['value'], var_uuid, i['secured'])
+                    bitbucket.manage_environment_variables('update', i['name'], i['value'], env_uuid, var_uuid, i['secured'])
                     result['changed'] = True
 
         else:
             # variable doesn't exist on current variables, add
-            bitbucket.manage_repository_variables('create', i['name'], i['value'], None, i['secured'])
+            bitbucket.manage_environment_variables('create', i['name'], i['value'], env_uuid, None, i['secured'])
             result['changed'] = True
 
     # action for current variables
@@ -134,12 +168,11 @@ def manage_variables(result, bitbucket, module):
                 for x in new_variables
         ):
             # current variable doesn't exists on new variables, delete
-            bitbucket.manage_repository_variables('delete', None, None, i['uuid'], None)
+            bitbucket.manage_environment_variables('delete', None, None, env_uuid, i['uuid'], None)
             result['changed'] = True
 
 def run_module():
     """ main module """
-
     # define available arguments/parameters a user can pass to the module
 
     variable_spec = dict(
@@ -162,13 +195,21 @@ def run_module():
         repository=dict(
             type='str',
             required=True,
+            no_log=False),
+        name=dict(
+            required=True,
+            type='str',
+            no_log=False,),
+        type=dict(
+            required=True,
+            type='str',
             no_log=False,
-            aliases=['name']),
+            choices=['Staging', 'Test', 'Production']),
         variables=dict(
-            type='list',
-            elements='dict',
             required=False,
             no_log=False,
+            type='list',
+            elements='dict',
             options=variable_spec),
     )
 
@@ -202,7 +243,7 @@ def run_module():
     existing_repository = bitbucket.get_repository_info()
 
     if existing_repository:
-        manage_variables(result, bitbucket, module)
+        manage_environments(result, bitbucket, module)
     else:
         module.fail_json(msg="Repository doesn't exists")
 
@@ -213,7 +254,7 @@ def run_module():
         module.exit_json(**result)
 
 def main():
-    """ main funtion """
+    """ main function """
 
     run_module()
