+++ src/tests/bitbucket-repo-var.yml
@@ -1,0 +1,23 @@
+- name: "Validate configuration of global variables on repo"
+  hosts: localhost
+  connection: local
+  gather_facts: false
+  become: false
+
+  vars_files:
+
+    - vars/vars.yml
+
+  tasks:
+
+    - name: Test repo_var module
+      i2btech.ops.bitbucket_repo_var:
+        username: "{{ bb_user }}"
+        password: "{{ bb_pass }}"
+        repository: "poc-sample"
+        variables:
+          - name: user
+            value: xxx
+          - name: pass
+            value: yyy
+            secured: true
