+++ src/tests/bitbucket-repo.yml
@@ -1,0 +1,19 @@
+- name: "Validate creation of repo"
+  hosts: localhost
+  connection: local
+  gather_facts: false
+  become: false
+
+  vars_files:
+
+    - vars/vars.yml
+
+  tasks:
+
+    - name: Test repo module
+      i2btech.ops.bitbucket_repo:
+        username: "{{ bb_user }}"
+        password: "{{ bb_pass }}"
+        repository: "poc-sample"
+        project_key: "POC"
+        state: "present"
