--- src/tests/sample-pb-repo-perm.yml
@@ -1,19 +1,0 @@
-- hosts: localhost
-  connection: local
-  gather_facts: false
-  become: false
-
-  tasks:
-
-    - name: Test repo_perm module
-      i2btech.ops.repo_perm:
-        username: "xxx"
-        password: "xxx"
-        repository: "poc-sample"
-        permissions:
-          - type: group
-            name: administrators
-            perm: admin
-          - type: group
-            name: developers
-            perm: write
