--- src/tests/sample-pb-repo-var.yml
@@ -1,18 +1,0 @@
-- hosts: localhost
-  connection: local
-  gather_facts: false
-  become: false
-
-  tasks:
-
-    - name: Test repo_var module
-      i2btech.ops.repo_var:
-        username: "xxx"
-        password: "xxx"
-        repository: "poc-sample"
-        variables:
-          - name: user
-            value: xxx
-          - name: pass
-            value: yyy
-            secured: True
