--- src/tests/sample-pb-repo.yml
@@ -1,14 +1,0 @@
-- hosts: localhost
-  connection: local
-  gather_facts: false
-  become: false
-
-  tasks:
-
-    - name: Test repo module
-      i2btech.ops.repo:
-        username: "xxx"
-        password: "xxx"
-        repository: "poc-sample"
-        project_key: "POC"
-        state: "present"
