+++ src/tests/vars/vars.yml
@@ -1,0 +1,3 @@
+---
+bb_user: "{{ lookup('ansible.builtin.env', 'BB_USER') }}"
+bb_pass: "{{ lookup('ansible.builtin.env', 'BB_PASS') }}"
